+++ app/backend/src/test/java/jp/onehr/oneops/OneOpsApplicationTests.java
@@ -23,7 +23,7 @@ class OneOpsApplicationTests {
     void setUp() {
         legacyGatewayProcess = mock(LegacyGatewayProcess.class);
         mockMvc = MockMvcBuilders
-            .standaloneSetup(new HealthController("0.18.16", legacyGatewayProcess))
+            .standaloneSetup(new HealthController("0.18.17", legacyGatewayProcess))
             .build();
     }